requirements.txt
python-dotenv
Flask
requests
langchain-google-genai
langchain